Improved validation mechanism
The validation mechanism has been reworked so that the tool can
support three levels of validation: 'none', 'warn', 'strict'. 'warn'
outputs the validation error to standard error and 'strict' does the
same but also exits the application. In both cases the offending
document is output as usual so that it can be checked manually.